pyproject.toml
[project]
name = "electricitymaps-airflow-scheduler"
description = "Electricity Maps' Airflow provider to schedule operations within an Airflow pipeline"
authors = [
    { name = "Pierre Segonne", email = "pierre.segonne@electricitymaps.com" },
]
license = { text = "Apache License 2.0" }
requires-python = ">=3.10,<3.13"
dependencies = ["apache-airflow (>=2.10.0,<4.0.0)"]
classifiers = [
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
]
dynamic = ["version"]

[tool.poetry]
name = "electricitymaps-airflow-scheduler"
version = "0.1.3"
description = "Electricity Maps' Airflow provider to schedule operations within an Airflow pipeline"
authors = [
    "Pierre Segonne <pierre.segonne@electricitymaps.com>",
    "Paul Romieu <paul.romieu@electricitymaps.com>",
    "Viktor Andersson <viktor.andersson@electricitymaps.com>" 
]
packages = [{ include = "electricitymaps_airflow_scheduler", from = "src" }]


[tool.poetry.group.test.dependencies]
pytest = "^8.4.2"


[tool.poetry.group.lint.dependencies]
ruff = "^0.13.0"


[tool.poetry.group.dev.dependencies]
graphviz = "^0.21"
pre-commit = "^4.0.0"

[project.urls]
Homepage = "https://www.electricitymaps.com/"
Source = "https://github.com/electricitymaps/electricitymaps-airflow-scheduler/"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.entry-points.apache_airflow_provider]
provider_info = "electricitymaps_airflow_scheduler:get_provider_info"

[tool.setuptools]
include-package-data = false

[tool.setuptools.dynamic]
version = { attr = "electricitymaps_airflow_scheduler.__version__" }

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["*tests.*", "*tests", ".ruff_cache", ".pytest_cache"]

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"
